Update 4600_export-utm-packages
The `plist` template has "Whonix-Gateway-Xfce" hard coded. This presents an issue when attempting to build Whonix-Gateway-CLI as the buildscript would create a raw image file with the filename `Whonix-Gateway-CLI.raw` while the the `config.plist`would point to a nonexistent `Whonix-Gateway-Xfce.raw`file. 

The same goes for the VM naming as it appears in UTM